[FIX] hr: make image zoomable again
The zoom feature of Image Widget expects a field `image_` being present
in the view to check if a zoomable image is available or not.

Since #69819 the image field was replaced by an avatar one thus breaking
the zoom.

This commit also puts the Chat icon on kanban cards of
hr.employee.public in the bottom right.

closes odoo/odoo#75099

Taskid: 2622671
X-original-commit: 0c7808d4e1d830e60a8a861d12c1dd1981a9f8a1
Signed-off-by: Kevin Baptiste <kba@odoo.com>