[IMP] website: add logged in param for conditional visibility
This commits adds to the existing conditional visibility option
introduced in [1], the option to display or hide blocs to a user,
depending on if he's logged in or not.

A select widget was added as an alternative to the many2one widgets that
were currently used, as we don't need to connect to the database to
display logged in/logged out option.

[1]: 9cd8d3c4b25ddd1e1509b9bed31343adc0448dd9

task-2618946

closes odoo/odoo#75107

Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>