[FIX] website_sale: fix cart popover position
On vertical menu with "flex-direction: column", cart item will take
full parent width, which makes the popover appear far from cart link
(in the middle of screen).

The goal of this commit is to fix this behaviour by adjusting cart
item width on large vertical menu.

task-2618929

Part-of: odoo/odoo#75063