[FIX]account,account_debit_note: not show only purchaseable product
Only purchaseable products are not searchable in the lines of debit notes
and credit notes if this account moves are created from their tree view with
the CREATE button. This changes make that this behavieour were also if we
create this account moves from the buttons in the invoices "ADD CREDIT NOTE"
and "ADD DEBIT NOTE".

closes odoo/odoo#78262

Task: https://www.odoo.com/my/task/2543493
X-original-commit: 8c22fe2ea6be27dc921d6c86213c79b31da8bf7b
Signed-off-by: Florian Gilbert <flg@odoo.com>
Signed-off-by: Olivier Colson <oco@odoo.com>