[FIX] hr_timesheet: add sudo to _apply_timesheet_label method
Before this commit, the portal user has no access to uom.uom model.
Therefore, this user has an access error when we want to see task form
view.

This commit adds a sudo to use this method to correctly display the
label for the unit_amount of timesheets shown in the task form view.

task-2379518

closes #73341